Restructure component folder structure
