+++ src/components/Stopwatch/_tests_/Stopwatch.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Enzyme from 'enzyme';
 import { shallow } from 'enzyme';
 import toJson from 'enzyme-to-json';
-import Stopwatch from './Stopwatch';
+import Stopwatch from '../Stopwatch';
 import Adapter from 'enzyme-adapter-react-16';
 
 Enzyme.configure({ adapter: new Adapter() });
