+++ src/components/Timer/_tests_/Timer.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Enzyme from 'enzyme';
-import Timer from './Timer';
+import Timer from '../Timer';
 import { shallow } from 'enzyme';
 import toJson from 'enzyme-to-json';
 import Adapter from 'enzyme-adapter-react-16';
